Streamlit/requirements.txt
streamlit>=1.30
numpy>=1.25
torch>=2.1
transformers>=4.40
huggingface-hub>=0.17
music21>=8.2
pytest>=8.0
google-genai>=0.2
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
